Refactored MangaTagger to MetadataManager
