+++ .gitignore
@@ -10,7 +10,7 @@ undo.json
 /ZIPPER_2_COVER_SETTER_UI.ui
 /junk_stuff/
 /MangaManager/CoverManagerLib/.pytest_cache/
-/MangaManager/MangaTaggerLib/.pytest_cache/
+/MangaManager/MetadataManagerLib/.pytest_cache/
 /MangaManager/tests/.pytest_cache/
 /MangaManager/tests/Sample CBZ Chapter 23.cbz
 /MangaManager/tests/Sample CBZ Chapter 24.cbz
