+++ MangaManager/MetadataManagerLib/ComicInfo.py
