+++ MangaManager/MetadataManagerLib/MetadataManager.py
