+++ MangaManager/MetadataManagerLib/cbz_handler.py
