+++ MangaManager/MetadataManagerLib/errors.py
